+++ src/dirsizes.rs
@@ -30,6 +30,7 @@ pub(crate) struct DirSizes<'a> {
     pub(crate) total_reg_cache_size: u64,      // registry cache size
     pub(crate) total_reg_src_size: u64,        // registry sources size
     pub(crate) total_reg_index_size: u64,      // registry index size
+    pub(crate) total_reg_index_num: u64,       // number of registry indices size
     pub(crate) numb_reg_cache_entries: usize,  // number of source archives
     pub(crate) numb_reg_src_checkouts: usize,  // number of source checkouts
     pub(crate) root_path: &'a std::path::PathBuf,
@@ -114,20 +115,21 @@ impl<'a> DirSizes<'a> {
 
         let total_size = total_reg_size + total_git_db_size + total_bin_size;
         Self {
-            total_size,                                      // total size of cargo root dir
-            numb_bins,                                       // number of binaries found
-            total_bin_size,                                  // total size of binaries found
-            total_reg_size,                                  // registry size
-            total_git_db_size,         // size of bare repos and checkouts combined
-            total_git_repos_bare_size, // git db size
-            numb_git_repos_bare_repos, // number of cloned repos
-            numb_git_checkouts,        // number of checked out repos
-            total_git_chk_size,        // git checkout size
-            total_reg_cache_size,      // registry cache size
-            total_reg_src_size,        // registry sources size
+            total_size,                           // total size of cargo root dir
+            numb_bins,                            // number of binaries found
+            total_bin_size,                       // total size of binaries found
+            total_reg_size,                       // registry size
+            total_git_db_size,                    // size of bare repos and checkouts combined
+            total_git_repos_bare_size,            // git db size
+            numb_git_repos_bare_repos,            // number of cloned repos
+            numb_git_checkouts,                   // number of checked out repos
+            total_git_chk_size,                   // git checkout size
+            total_reg_cache_size,                 // registry cache size
+            total_reg_src_size,                   // registry sources size
             total_reg_index_size: reg_index_size, // registry index size
+            total_reg_index_num: registry_index_cache.number_of_indices(), // number of indices //@TODO parallelize like the rest
             numb_reg_cache_entries: total_reg_cache_entries, // number of source archives
-            numb_reg_src_checkouts,    // number of source checkouts
+            numb_reg_src_checkouts,                          // number of source checkouts
             root_path,
         }
     }
@@ -176,19 +178,36 @@ impl<'a> fmt::Display for DirSizes<'a> {
             )
         )?;
 
-        write!(
-            f,
-            "{}",
-            library::pad_strings(
-                0,
-                45,
-                "    Registry index: ",
-                &self
-                    .total_reg_index_size
-                    .file_size(file_size_opts::DECIMAL)
-                    .unwrap(),
-            )
-        )?;
+        // how many indices do we have?
+        if self.total_reg_index_num == 1 {
+            write!(
+                f,
+                "{}",
+                library::pad_strings(
+                    0,
+                    45,
+                    "    Registry index: ",
+                    &self
+                        .total_reg_index_size
+                        .file_size(file_size_opts::DECIMAL)
+                        .unwrap(),
+                )
+            )?;
+        } else {
+            write!(
+                f,
+                "{}",
+                library::pad_strings(
+                    0,
+                    45,
+                    &format!("    {} registry indices: ", &self.total_reg_index_num),
+                    &self
+                        .total_reg_index_size
+                        .file_size(file_size_opts::DECIMAL)
+                        .unwrap(),
+                )
+            )?;
+        }
 
         write!(
             f,
@@ -316,6 +335,7 @@ mod libtests {
                 numb_reg_src_checkouts: reg_src.file_number as usize,
 
                 total_reg_index_size: reg_index.dir_size,
+                total_reg_index_num: 1,
                 root_path: path,
             }
         }