+++ tests/alternative_registry.rs
@@ -251,7 +251,7 @@ rayon = { version = "1", registry = "my-index" }
     Total:                              219.45 MB
       0 installed binaries:                  0  B
       Registry:                         219.45 MB
-        Registry index:                 211.80 MB
+        2 registry indices:                 211.80 MB
         22 crate archives:                1.39 MB
         22 crate source checkouts:        6.25 MB
       Git db:                                0  B
@@ -262,10 +262,10 @@ rayon = { version = "1", registry = "my-index" }
     desired_output.push_str(
         "Total:                              .* MB
   0 installed binaries:                  0  B
-  Registry:                         .* MB
-    Registry index:                 .* MB
-    .. crate archives:                .* MB
-    .. crate source checkouts:        .* MB
+  Registry:                        .* MB
+    2 registry indices:            .* MB
+    .. crate archives:             .* MB
+    .. crate source checkouts:     .* MB
   Git db:                                0  B
     0 bare git repos:                    0  B
     0 git repo checkouts:                0  B",
@@ -275,7 +275,7 @@ rayon = { version = "1", registry = "my-index" }
 
     assert!(
         regex.clone().is_match(&stdout),
-        "regex: {:?}, cc_output: {}",
+        "ERROR: regex did not match!\n\nregex:\n{:?}\n\ncc_output:\n{:?}",
         regex,
         stdout
     );